AGV cleanup and floating point double fix

+++ Database/Models/CalculationResultModel.cs
@@ -7,9 +7,9 @@ public class CalculationResultModel
     public Guid SegmentId { get; set; }
     public int SegmentNo { get; set; }
     public int StatorNo { get; set; }
-    public double MeasuredValue { get; set; }
-    public double Tolerance { get; set; }
-    public double? Deviation { get; set; }
+    public decimal MeasuredValue { get; set; }
+    public decimal Tolerance { get; set; }
+    public decimal? Deviation { get; set; }
     public DateTime Date { get; set; }
     public bool? Adjustment { get; set; }
 }
