+++ Shared/Models/AdjustedCalculationDto.cs
@@ -3,9 +3,9 @@
 public class AdjustedCalculationDto
 {
     public DateTime Date { get; set; }
-    public double MeasuredValue { get; set; }
-    public double Tolerance { get; set; }
-    public double? Deviation { get; set; }
+    public decimal MeasuredValue { get; set; }
+    public decimal Tolerance { get; set; }
+    public decimal? Deviation { get; set; }
     public bool? AdjustmentNeeded { get; set; }
     public int SegmentNo { get; set; }
     public int? StatorNo { get; set; }
